+++ src/org/capitalism/TransactionTerms.java
@@ -1,14 +1,16 @@
 package org.capitalism;
 
+import java.math.BigDecimal;
+
 public class TransactionTerms {
 
-	public TransactionTerms(double value, int consumableId) {
+	public TransactionTerms(BigDecimal value, int consumableId) {
 
 		this.transactionValue = value;
 		this.consumableId = consumableId;
 	}
 
-	final public double transactionValue;
+	final public BigDecimal transactionValue;
 	final public int consumableId;
 
 }
